ES GEHT

+++ src/Wheat.java
@@ -1,9 +1,9 @@
 public class Wheat {
 
-    int numberOfGrain;
+    short numberOfGrain;
 
 
     public Wheat() {
-        this.numberOfGrain = 100 + (int)(Math.random() * ((500 - 100) + 1));
+        this.numberOfGrain = (short) (100 + (int)(Math.random() * ((500 - 100) + 1)));
     }
 }
